refactor: restructure files and libraries
