+++ lib/pages/splash/splash_screen_app.dart
